Adding functions to query purchases/consumption within a given date range. Refactoring into model.go, query.go, and update.go files.
